refactor: Separate Clients and Policies
