+++ src/Dodo.HttpClient.ResiliencePolicies/HttpClientBuilderExtensionsPolicies.cs
@@ -1,7 +1,5 @@
-using System;
 using System.Net;
 using System.Net.Http;
-using System.Net.Http.Headers;
 using Dodo.HttpClientResiliencePolicies.CircuitBreakerPolicy;
 using Dodo.HttpClientResiliencePolicies.RetryPolicy;
 using Dodo.HttpClientResiliencePolicies.TimeoutPolicy;
@@ -17,57 +15,8 @@ namespace Dodo.HttpClientResiliencePolicies
 	/// <summary>
 	/// Extension methods for configuring <see cref="IHttpClientBuilder"/> with Polly retry, timeout, circuit breaker policies.
 	/// </summary>
-	public static class HttpClientBuilderExtensions
+	public static class HttpClientBuilderExtensionsPolicies
 	{
-		/// <summary>
-		/// Adds the <see cref="IHttpClientFactory"/> and related services to the <see cref="IServiceCollection"/>
-		/// with pre-configured JSON headers, client Timeout and default policies.
-		/// </summary>
-		/// <returns>An <see cref="IHttpClientBuilder"/> that can be used to configure the client.</returns>
-		public static IHttpClientBuilder AddJsonClient<TClientInterface, TClientImplementation>(
-			this IServiceCollection sc,
-			Uri baseAddress,
-			string clientName = null) where TClientInterface : class
-			where TClientImplementation : class, TClientInterface
-		{
-			return AddJsonClient<TClientInterface, TClientImplementation>(
-				sc, baseAddress, new ResiliencePoliciesSettings(), clientName);
-		}
-
-		/// <summary>
-		/// Adds the <see cref="IHttpClientFactory"/> and related services to the <see cref="IServiceCollection"/>
-		/// with pre-configured JSON headers, client Timeout and default policies.
-		/// </summary>
-		/// <returns>An <see cref="IHttpClientBuilder"/> that can be used to configure the client.</returns>
-		public static IHttpClientBuilder AddJsonClient<TClientInterface, TClientImplementation>(
-			this IServiceCollection sc,
-			Uri baseAddress,
-			ResiliencePoliciesSettings settings,
-			string clientName = null) where TClientInterface : class
-			where TClientImplementation : class, TClientInterface
-		{
-			var delta = TimeSpan.FromMilliseconds(1000);
-
-			void DefaultClient(HttpClient client)
-			{
-				client.BaseAddress = baseAddress;
-				client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
-				client.Timeout = settings.OverallTimeoutPolicySettings.Timeout + delta;
-			}
-
-			var httpClientBuilder = string.IsNullOrEmpty(clientName)
-				? sc.AddHttpClient<TClientInterface, TClientImplementation>(DefaultClient)
-				: sc.AddHttpClient<TClientInterface, TClientImplementation>(clientName, DefaultClient);
-
-			httpClientBuilder
-				.AddTimeoutPolicy(settings.OverallTimeoutPolicySettings)
-				.AddRetryPolicy(settings.RetrySettings)
-				.AddCircuitBreakerPolicy(settings.CircuitBreakerSettings)
-				.AddTimeoutPolicy(settings.TimeoutPerTryPolicySettings);
-
-			return httpClientBuilder;
-		}
-
 		/// <summary>
 		/// Adds pre-configured resilience policies.
 		/// </summary>
